Save game back-compatibility
When loading a save game that did not have Archite Reinforcement enabled, the collections storing the levels for stat and capacity upgrades would be Null. This causes an issue, because every part of the codebase depends on these fields never being null. The solution is to run a null-check whenever the tracker is loaded, then initialize the upgrade collections when necessary.

- Add a null check to the tracker's scribe function for the upgrade level collections.