Initial NPC Archite generation
Add support for NPCs to spawn with archites.

Currently, this just grants a large quantity of unspent upgrade points to the pawn. Point allocation logic still needs to be implemented. Additionally, all non-player-faction pawns have a 100% chance of generating with archites right now for testing purposes, so actual generation probabilities will need to be implemented soon.