Corpse reclamation value fix
The number of archites reclaimed from a corpse is now based on the cumulative number of archites instead of the number of upgrade points.

The unclear distinction between archites, upgrade progress, and upgrade points has created a substantial amount of technical debt, and I intend to refactor the naming convention of CompArchiteTracker's members once I start developing prerelease 5.